% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingStations.R
\name{check_missing_stations}
\alias{check_missing_stations}
\title{Checks for Missing Stations}
\usage{
check_missing_stations(con)
}
\arguments{
\item{con}{The name of the created database.}
}
\value{
A list of any stations that are missing in the created database.
}
\description{
Compares the database station list with the hourly climate station list.
}
