% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissingTime.R
\name{get_missing_over_time}
\alias{get_missing_over_time}
\title{Calculates Missing Data Over Time}
\usage{
get_missing_over_time(con, column)
}
\arguments{
\item{con}{The name of the created database.}

\item{column}{Character string. Created by the user to specify the column
of interest (e.g., \code{"Temp_C"}).}
}
\value{
A data frame with:
\itemize{
  \item Year
  \item Month
  \item total_obs
  \item missing_count
  \item pct_missing
  \item time_label
}
}
\description{
Calculates the percentage of missing values for a selected variable
grouped by year and month.Can be used to plot.
}
