% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatioDataMissing.R
\name{get_missing_ratio}
\alias{get_missing_ratio}
\title{Calculates Missing Data Ratios for Selected Variables}
\usage{
get_missing_ratio(con, columns)
}
\arguments{
\item{con}{The name of the created database.}

\item{columns}{Character vector. Created by the user to specify the column
of interest (e.g., \code{c("Temp_C", "Rel_Hum")}).}
}
\value{
A one-row data frame containing missing value proportions.
}
\description{
Calculates the proportion of missing values (NA/NULL) for selected variables.
}
