% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationRanges.R
\name{get_observation_ranges}
\alias{get_observation_ranges}
\title{Get Observation Time Ranges by Station}
\usage{
get_observation_ranges(con)
}
\arguments{
\item{con}{The name of the created database.}
}
\value{
A data frame containing:
\itemize{
  \item Climate_ID
  \item start_date
  \item end_date
  \item n_obs
}
}
\description{
Shows the time span and total number of observations for each station.
}
