% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepeatedObds.R
\name{get_repeated_observations}
\alias{get_repeated_observations}
\title{Identify Consecutive Repeated Observations}
\usage{
get_repeated_observations(con, column)
}
\arguments{
\item{con}{The name of the created database.}

\item{column}{Character string. Created by the user to specify the column
of interest (e.g., \code{"Temp_C"}).}
}
\value{
A data frame of rows where consecutive values are repeated.
}
\description{
Detects consecutive repeated values for a user specified variable
}
