% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataEmbed.R
\name{get_station_file}
\alias{get_station_file}
\title{Accesses embedded CSV}
\usage{
get_station_file()
}
\description{
Accesses embedded CSV
}
\details{
Used by the other functions, does not need to be run by the user.
}
