% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkippedObs.R
\name{get_time_gaps}
\alias{get_time_gaps}
\title{Identifies Missing Hourly Time Gaps}
\usage{
get_time_gaps(con)
}
\arguments{
\item{con}{The name of the created database.}
}
\value{
A data frame containing observations where time gaps exceed one hour.
}
\description{
Checks for time gaps greater than one hour between consecutive
observations.
}
